fix: moved things around
