+++ aws_lambda_powertools/event_handler/openapi/exceptions.py
@@ -2,6 +2,10 @@ from typing import Any, Sequence
 
 
 class ValidationException(Exception):
+    """
+    Base exception for all validation errors
+    """
+
     def __init__(self, errors: Sequence[Any]) -> None:
         self._errors = errors
 
@@ -10,6 +14,10 @@ class ValidationException(Exception):
 
 
 class RequestValidationError(ValidationException):
+    """
+    Raised when the request body does not match the OpenAPI schema
+    """
+
     def __init__(self, errors: Sequence[Any], *, body: Any = None) -> None:
         super().__init__(errors)
         self.body = body
